pyproject.toml
[project]
name = "531weighted"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "PyYAML>=6.0.2",
    "ruff>=0.15.7",
    "streamlit>=1.55.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
